% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dividend_yield.R
\name{get_divdend_yield}
\alias{get_divdend_yield}
\title{Download Dividend Yield}
\usage{
get_divdend_yield(agg = "daily", agg.func = NULL, progress = TRUE)
}
\arguments{
\item{agg}{Frequency to aggregate. It can be one of 'year' (or 'yearly') and ('month' or 'monthly'). If it is NULL or 'day' (or 'daily'), then it is the raw data from the website.}

\item{agg.func}{A string that represents a function to aggregate. It can be any function that dplyr's `summarise()` function handles. If missing, then `last()` is used by deafult.}

\item{progress}{If it is TRUE, then it shows download progress.}
}
\value{
A dataframe with the data aggregated (or not)
}
\description{
Downloads and aggregates dividend yield data from NEFIN's website.
The user can also aggregate data in a daily (default and how raw data is provided), monthly and yearly fashion.
}
\examples{

agg <- 'month'
agg.func <- 'last'

\dontrun{
df.out <- get_divdend_yield(agg = agg, agg.func = agg.func)
}
}
