% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_illiquidity_index.R
\name{get_illiquidity_index}
\alias{get_illiquidity_index}
\title{Download Market Illiquidity Index}
\usage{
get_illiquidity_index(agg = "month", agg.func = NULL, progress = TRUE)
}
\arguments{
\item{agg}{Frequency to aggregate. It can be year' (or 'yearly'). If it is NULL or 'month' (or 'monthly') then it is the raw data from the website.}

\item{agg.func}{A string that represents a function to aggregate. It can be any function that dplyr's `summarise()` function handles. If missing, then `last()` is used by deafult.}

\item{progress}{If it is TRUE, then it shows download progress.}
}
\value{
A dataframe with the data aggregated (or not)
}
\description{
Downloads and aggregates Market Illiquidity Index data from NEFIN's website.
The user can also aggregate data in a monthly (default) and yearly fashion.
}
\examples{

agg <- 'month'
agg.func <- 'last'

\dontrun{
df.out <- get_illiquidity_index(agg = agg, agg.func = agg.func)
}
}
