% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_risk_factors.R
\name{get_risk_factors}
\alias{get_risk_factors}
\title{Download One or More Risk Factors}
\usage{
get_risk_factors(factors, agg = "month", agg.func = "last", progress = TRUE)
}
\arguments{
\item{factors}{Risk factors to download. If it is 'all', then all factors are concatenated. It can be one or more of the available factors from NEFIN's website: 'Market', 'SMB', 'HML', 'WML', 'IML', 'Rf' (Risk-free).}

\item{agg}{Frequency to aggregate. It can be one of 'year' (or 'yearly') and ('month' or 'monthly'). If it is NULL or 'day' (or 'daily'), then it is the raw data from the website.}

\item{agg.func}{A string that represents a function to aggregate. It can be any function that dplyr's `summarise()` function handles.}

\item{progress}{If it is TRUE, then it shows download progress.}
}
\value{
A dataframe with the data aggregated (or not)
}
\description{
Downloads and aggregates a chosen risk factors data from NEFIN's website. If more than one factor is provided, then the resulting data frame is the full join of single factor's dataframe.
The user can also aggregate data in a daily (default and how raw data is provided), monthly and yearly fashion.
}
\examples{

agg <- 'month'
agg.func <- 'last'

df.out <- get_risk_factors('all', agg = agg, agg.func = agg.func)
print(head(df.out))
}
