% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sector_cost_of_capital.R
\name{get_sector_cost_of_capital}
\alias{get_sector_cost_of_capital}
\title{Download Cost of Capital}
\usage{
get_sector_cost_of_capital(
  sector,
  agg = NULL,
  agg.func = NULL,
  progress = TRUE
)
}
\arguments{
\item{sector}{Sector to download data. It can be one of the available sectors from NEFIN's website: ''Basic', 'Construction', 'Consumer', 'Energy', 'Finance', 'Manufacturing' and 'Other'.}

\item{agg}{Frequency to aggregate. It can be one of 'year' (or 'yearly'). If it is NULL or 'month' (or 'monthly'), then it is the raw data from the website.}

\item{agg.func}{A string that represents a function to aggregate. It can be any function that dplyr's `summarise()` function handles. If missing, then `last()` is used by deafult.}

\item{progress}{If it is TRUE, then it shows download progress.}
}
\value{
A dataframe with the data aggregated (or not)
}
\description{
Downloads and aggregates cost of capital of a given industrial sector, from NEFIN's website.
The user can also aggregate data in a monthly (default and how raw data is provided) and yearly fashion.
}
\examples{

my.sector <- 'Consumer'
agg <- 'year'
agg.func <- 'last'

\dontrun{
df.out <- get_sector_cost_of_capital(my.sector, agg = agg, agg.func = agg.func)
}
}
