% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_risk_factor.R
\name{get_single_risk_factor}
\alias{get_single_risk_factor}
\title{Download Risk Factor}
\usage{
get_single_risk_factor(
  factor,
  agg = "daily",
  agg.func = "last",
  progress = TRUE
)
}
\arguments{
\item{factor}{Risk factor to download. It can be one of the available factors from NEFIN's website: 'Market', 'SMB', 'HML', 'WML', 'IML', 'Rf' (Risk-free).}

\item{agg}{Frequency to aggregate. It can be one of 'year' (or 'yearly') and ('month' or 'monthly'). If it is NULL or 'day' (or 'daily'), then it is the raw data from the website.}

\item{agg.func}{A string that represents a function to aggregate. It can be any function that dplyr's `summarise()` function handles. If missing, then `last()` is used by deafult.}

\item{progress}{If it is TRUE, then it shows download progress.}
}
\value{
A dataframe with the data aggregated (or not)
}
\description{
Downloads and aggregates a chosen risk factor data from NEFIN's website.
The user can also aggregate data in a daily (default and how raw data is provided), monthly and yearly fashion.
}
\examples{

my.factor <- 'Market'
agg <- 'month'
agg.func <- 'last'

\dontrun{
df.out <- get_single_risk_factor(my.factor, agg = agg, agg.func = agg.func)
}
}
